refactor(openmeteo): split current condition parser
Move Open-Meteo current condition parsing, precipitation inference, and ISO timestamp handling into a focused module while preserving weather_client_openmeteo exports.

Tested: pytest tests/test_weather_client_openmeteo_parser.py tests/test_openmeteo_forecast_timezone.py tests/test_openmeteo_client.py tests/test_weather_client.py tests/test_weather_client_trends.py; ruff check src tests scripts; pyright.

Confidence: high
